Working Redirect for ACME Challenge, Removing TestApplication
For setting up the Let's Encrypt, we needed a specific redirect rule for ACME Challenges.
The challenge was being put in www.wispymouse.net, instead of proma.wispymouse.net.
It's unknown if this was due to rewrite rules that were in place, but this specific failsafe helped it pass the challenge.

TestApplication was also removed, and proma.wispymouse.net is now a subdomain.